+++ R/phishnet_get_show_rating.R
@@ -3,7 +3,6 @@
 #' @importFrom jsonlite fromJSON
 #' @importFrom httr GET
 #' @import jsonlite
-#' @importFrom textreadr read_html
 #' @importFrom dplyr filter
 #' @importFrom stringr str_split_fixed
 #' @importFrom zoo na.locf