+++ tests/testthat/test-phishnet.R
@@ -1,0 +1,27 @@
+test_that('pn_get_all_showdates is working', {
+
+})
+
+test_that('pn_get_all_songs is working', {
+
+})
+
+test_that('pn_get_all_setlists is working', {
+
+})
+
+test_that('pn_get_all_venues is working', {
+
+})
+
+test_that('pn_get_setlist is working', {
+
+})
+
+test_that('pn_get_showrating is working', {
+
+})
+
+test_that('pn_get_shownotes is working', {
+
+})