+++ NEWS.md
@@ -1,3 +1,8 @@
 # phishr 0.1.0.9000
 
 * Added _extremely_ experimental connection to phish.in API!
+
+# phishr 1.0.0 
+
+* added V5 phish.net wrapper functions
+* added suite of unit tests